requirements.txt
# NutriVision - Requirements
# Python 3.8+

# Core ML/DL
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
pillow>=10.0.0

# Web Application
streamlit>=1.25.0
altair>=5.0.0

# Optional: PDF export (if implementing PDF feature)
# fpdf2>=2.7.0

# Optional: Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
